[adone] refactoring, additions * [collections] decomposition, add MapCache, NSCache, RefcountedCache, Set * [data] add protobuf, base32, basex, base58, base64url, varint, varintSigned * [is] improve function and asyncFunction, add emitter and asyncEmitter * add fake
