+++ types/adone/glosses/collections/index.d.ts
@@ -1,0 +1,28 @@
+/// <reference path="./array_set.d.ts" />
+/// <reference path="./async_queue.d.ts" />
+/// <reference path="./avl_tree.d.ts" />
+/// <reference path="./binary_search_tree.d.ts" />
+/// <reference path="./buffer_list.d.ts" />
+/// <reference path="./byte_array.d.ts" />
+/// <reference path="./default_map.d.ts" />
+/// <reference path="./fast_lru.d.ts" />
+/// <reference path="./linked_list.d.ts" />
+/// <reference path="./lru.d.ts" />
+/// <reference path="./map_cache.d.ts" />
+/// <reference path="./ns_cache.d.ts" />
+/// <reference path="./priority_queue.d.ts" />
+/// <reference path="./queue.d.ts" />
+/// <reference path="./rb_tree.d.ts" />
+/// <reference path="./refcounted_cache.d.ts" />
+/// <reference path="./set.d.ts" />
+/// <reference path="./stack.d.ts" />
+/// <reference path="./timedout_map.d.ts" />
+
+declare namespace adone {
+    /**
+     * Data structures
+     */
+    namespace collection {
+        //
+    }
+}
