+++ types/adone/glosses/collections/refcounted_cache.d.ts
@@ -1,0 +1,9 @@
+declare namespace adone.collection {
+    class RefcountedCache<T = any> extends MapCache<T> {
+        ref(key: string): void;
+
+        unref(key: string): void;
+
+        references(key: string): number;
+    }
+}
