+++ types/adone/index.d.ts
@@ -8,12 +8,13 @@
 /// <reference path="./glosses/application.d.ts" />
 /// <reference path="./glosses/archives.d.ts" />
 /// <reference path="./glosses/assertion.d.ts" />
-/// <reference path="./glosses/collections.d.ts" />
+/// <reference path="./glosses/collections/index.d.ts" />
 /// <reference path="./glosses/compressors.d.ts" />
 /// <reference path="./glosses/data.d.ts" />
 /// <reference path="./glosses/datetime.d.ts" />
 /// <reference path="./glosses/events.d.ts" />
 /// <reference path="./glosses/exceptions.d.ts" />
+/// <reference path="./glosses/fake.d.ts" />
 /// <reference path="./glosses/fast.d.ts" />
 /// <reference path="./glosses/fs.d.ts" />
 /// <reference path="./glosses/is.d.ts" />
