+++ types/adone/tsconfig.json
@@ -22,6 +22,7 @@
     "files": [
         "index.d.ts",
         "adone.d.ts",
+        "glosses/fake.d.ts",
         "glosses/math/index.d.ts",
         "glosses/math/matrix.d.ts",
         "glosses/math/simd.d.ts",
@@ -32,7 +33,26 @@
         "glosses/promise.d.ts",
         "glosses/shani.d.ts",
         "glosses/shani-global.d.ts",
-        "glosses/collections.d.ts",
+        "glosses/collections/index.d.ts",
+        "glosses/collections/array_set.d.ts",
+        "glosses/collections/async_queue.d.ts",
+        "glosses/collections/avl_tree.d.ts",
+        "glosses/collections/binary_search_tree.d.ts",
+        "glosses/collections/buffer_list.d.ts",
+        "glosses/collections/byte_array.d.ts",
+        "glosses/collections/default_map.d.ts",
+        "glosses/collections/fast_lru.d.ts",
+        "glosses/collections/linked_list.d.ts",
+        "glosses/collections/lru.d.ts",
+        "glosses/collections/map_cache.d.ts",
+        "glosses/collections/ns_cache.d.ts",
+        "glosses/collections/priority_queue.d.ts",
+        "glosses/collections/queue.d.ts",
+        "glosses/collections/rb_tree.d.ts",
+        "glosses/collections/refcounted_cache.d.ts",
+        "glosses/collections/set.d.ts",
+        "glosses/collections/stack.d.ts",
+        "glosses/collections/timedout_map.d.ts",
         "glosses/events.d.ts",
         "glosses/fs.d.ts",
         "glosses/is.d.ts",
@@ -46,7 +66,27 @@
         "adone-tests.ts",
         "test/index.ts",
         "test/index-import.ts",
+        "test/glosses/fake.ts",
         "test/glosses/application.ts",
+        "test/glosses/collections/array_set.ts",
+        "test/glosses/collections/async_queue.ts",
+        "test/glosses/collections/avl_tree.ts",
+        "test/glosses/collections/binary_search_tree.ts",
+        "test/glosses/collections/buffer_list.ts",
+        "test/glosses/collections/byte_array.ts",
+        "test/glosses/collections/default_map.ts",
+        "test/glosses/collections/fast_lru.ts",
+        "test/glosses/collections/linked_list.ts",
+        "test/glosses/collections/lru.ts",
+        "test/glosses/collections/map_cache.ts",
+        "test/glosses/collections/ns_cache.ts",
+        "test/glosses/collections/priority_queue.ts",
+        "test/glosses/collections/queue.ts",
+        "test/glosses/collections/rb_tree.ts",
+        "test/glosses/collections/refcounted_cache.ts",
+        "test/glosses/collections/set.ts",
+        "test/glosses/collections/stack.ts",
+        "test/glosses/collections/timedout_map.ts",
         "test/glosses/math/index.ts",
         "test/glosses/math/matrix.ts",
         "test/glosses/math/simd.ts",
@@ -57,7 +97,6 @@
         "test/glosses/promise.ts",
         "test/glosses/shani.ts",
         "test/glosses/shani-global.ts",
-        "test/glosses/collections.ts",
         "test/glosses/events.ts",
         "test/glosses/fs.ts",
         "test/glosses/is.ts",
